\name{bp2.find.proteins}
\alias{bp2.find.proteins}
\title{Find UniprotIDs in the list of physical entities}
\description{Internal function. }
\usage{
bp2.find.proteins(physical.entities)
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
  \item{physical.entities}{List of bp2.physicalEntity objects}

}

\value{
Character vector of UniprotIDs found wihtin physical.entities.
}

