\name{bp2.get.all.class.names}
\alias{bp2.get.all.class.names}
\title{List BioPAX model class names}
\description{List all class names in bp2 model.}
\usage{
bp2.get.all.class.names()
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
  \item{arg}{
  }

}

\value{
Character vector containg all class names of all BioPAX level 2 classes known to bp2.
}

\seealso{bp2.create.model}

\examples{
my.bp2 <- bp2.create.model(...)
bp2.get.all.class.names()
}
