\name{bp2.get.all.slot.names}
\alias{bp2.get.all.slot.names}
\title{List all slot names that are used in bp2 model objects}
\description{List all slot names that are used in bp2 model objects}
\usage{
bp2.get.all.slot.names()
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
}

\value{
Character vector of slot names.
}

\examples{
my.bp2 <- bp2.create.model(...)
bp2.get.all.slot.names()
}
