\name{bp2.get.uniprot.ids}
\alias{bp2.get.uniprot.ids}
\title{Get unitprot ID of a protein object}
\description{Returns UnioprotID character string of "bp2.protein" object}
\usage{
bp2.get.uniprot.ids(proteins)
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
  \item{proteins}{list of "bp2.protein" objects}
}

\value{
Vector of Uniprot ID strings of \code{proteins}
}

\examples{

}
