\name{bp2.ids2objects}
\alias{bp2.ids2objects}
\title{Convert numeric identifiers of bp2 model objects to S4 objects}
\description{Internal function. }
\usage{
bp2.ids2objects(ids)
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
  \item{ids}{Vector of numeric ids of bp2.model objects}
}

\seealso{\code{\link{bp2.ids2objects}}, \code{\link{bp2.resume}}}

\value{
List of objects retrieved from the active bp2.model.
}

\examples{
}
