\name{bp.objects2ids}
\alias{bp2.objects2ids}
\title{Convert bp2 model objects to numeric identifiers}
\description{Internal function. }
\usage{
bp2.objects2ids(objs)
}

\author{Ossi Koivistoinen}
\note{}
\arguments{
  \item{objs}{List of objects that inherit \code{bp2.rdf.entity}}
  }

\value{
Vector of integers containing ids of \code{objs}
}

\seealso{\code{\link{bp2..ids2objects}}, \code{\link{bp2.create.model}}}

\examples{
}
